create function CRUD user data

+++ app/Controllers/Users.php
@@ -45,6 +45,9 @@ class Users extends BaseController
 	}
 	public function deleteRole($role)
 	{
+		if (!$role) {
+			return redirect()->to(base_url('users'));
+		}
 		$deleteRole = $this->userModel->deleteRole($role);
 		if ($deleteRole) {
 			session()->setFlashdata('notif_success', '<b>Successfully added menu data</b> ');
@@ -62,6 +65,46 @@ class Users extends BaseController
 			return redirect()->to(base_url('users'));
 		} else {
 			session()->setFlashdata('notif_error', '<b>Failed to add menu data</b> ');
+			return redirect()->to(base_url('users'));
+		}
+	}
+	public function createUser()
+	{
+		if (!$this->validate(['inputUsername' => ['rules' => 'is_unique[users.username]']])) {
+			session()->setFlashdata('notif_error', '<b>Failed to add new user</b> The user already exists! ');
+			return redirect()->to(base_url('users'));
+		}
+		$createUser = $this->userModel->createUser($this->request->getPost(null, FILTER_SANITIZE_STRING));
+		if ($createUser) {
+			session()->setFlashdata('notif_success', '<b>Successfully added new user</b> ');
+			return redirect()->to(base_url('users'));
+		} else {
+			session()->setFlashdata('notif_error', '<b>Failed to add new user</b> ');
+			return redirect()->to(base_url('users'));
+		}
+	}
+	public function updateUser()
+	{
+		$updateUser = $this->userModel->updateUser($this->request->getPost(null, FILTER_SANITIZE_STRING));
+		if ($updateUser) {
+			session()->setFlashdata('notif_success', '<b>Successfully update user data</b> ');
+			return redirect()->to(base_url('users'));
+		} else {
+			session()->setFlashdata('notif_error', '<b>Failed to update user data</b> ');
+			return redirect()->to(base_url('users'));
+		}
+	}
+	public function deleteUser($userID)
+	{
+		if (!$userID) {
+			return redirect()->to(base_url('users'));
+		}
+		$deleteUser = $this->userModel->deleteUser($userID);
+		if ($deleteUser) {
+			session()->setFlashdata('notif_success', '<b>Successfully delete user</b> ');
+			return redirect()->to(base_url('users'));
+		} else {
+			session()->setFlashdata('notif_error', '<b>Failed to delete user</b> ');
 			return redirect()->to(base_url('users'));
 		}
 	}
